% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flexfile.R
\name{apply_flexfile}
\alias{apply_flexfile}
\title{Apply a function over a list of flexfiles}
\usage{
apply_flexfile(flexfile, fun, ...)
}
\arguments{
\item{flexfile}{A single flexfile or a list of flexfiles imported through the \code{read_flexfile} function.}

\item{fun}{The function to be applied.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
The original structure with the function applied.
}
\description{
\code{apply_flexfile()} is used to apply a function over each element of a list of
objects of type 'flexfile'. If the input is a single 'flexfile' is (i.e., not a list),
then the function will be applied only to that item.\cr
\cr
Therefore, use \code{apply_flexfile} when you wish to apply a function and are unsure
if the input is a list of flexfiles or a single flexfile.
}
