% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantityreport-class.R
\name{as_quantityreport}
\alias{as_quantityreport}
\title{Create object of type quantityreport}
\usage{
as_quantityreport(x, .show_check = TRUE)
}
\description{
Create object of type quantityreport
}
