% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model_to_snake.R
\name{change_case}
\alias{change_case}
\title{Convert table names}
\usage{
change_case(table_list, table_spec, from_case = NULL, to_case = NULL)
}
\arguments{
\item{table_list}{A list of tables to rename.}

\item{table_spec}{Data model specification object.}

\item{from_case}{String name of the case the object originates from. If \code{NULL},
defaults to the base case.}

\item{to_case}{String name of the case to convert to. If \code{NULL},
defaults to the base case.}
}
\description{
\code{change_case()} renames from one case to another from the file spec. \cr
\cr
Note that this function set may move to another location in a future version.
}
\details{
This function relies on names as set from a spec file. In order for this to work
consistently, a few things are assumed.\cr
\cr
The base case (or original/native case) tables are specified by the column named
'table' in both the 'tables' and 'fields' data frames of the spec. The fields
for the base case are specified by the column named 'field' in the 'fields data
frame of the spec.\cr
\cr
Any other valid case (e.g., 'new_case') specifies the table name in the 'tables' data
frame with the suffix "_table" (e.g., 'new_case_table'). The field names are specified in
the 'fields' table with the suffix "_name" (e.g., 'new_case_name').\cr
\cr
To access this specific case for conversion, simple use 'new_case' as the argument value for
either \code{from_case} or \code{to_case}.
}
