% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexfile_family.R
\name{create_flexfile_family}
\alias{create_flexfile_family}
\title{Create a FlexFile Family}
\usage{
create_flexfile_family(..., report_list = NULL)
}
\arguments{
\item{...}{One or more reports in the FlexFile family of types.}

\item{report_list}{Alternatively, a list of reports. If provided then the
contents of \code{...} will be ignored.}
}
\value{
A list of reports of type \code{flexfile_family}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{create_flexfile_family()} creates a list of related FlexFile Family reports. For example,
it is used to store the matching FlexFile and Quantity Data Report together.
}
