% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model_to_snake.R
\name{data_model_to_snake}
\alias{data_model_to_snake}
\alias{snake_to_data_model}
\title{Convert table names}
\usage{
data_model_to_snake(table_list, table_spec)

snake_to_data_model(table_list, table_spec)
}
\arguments{
\item{table_list}{A list of tables to rename.}

\item{table_spec}{Data model specification object.}
}
\description{
\code{data_model_to_snake()} renames from the data model notation to snake_case. This
function will also add any missing columns from the spec to the model. \cr
\cr
Note that this function set may move to another location in a future version.

\code{snake_to_data_model()} renames from snake_case to data model notation.
}
