% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{duplicated_report}
\alias{duplicated_report}
\title{Check for duplication in the metadata}
\usage{
duplicated_report(x)
}
\arguments{
\item{x}{A flexfile or quantityreport}
}
\description{
Check for duplication in the metadata
}
\keyword{internal}
