% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-class.R
\name{flat_class}
\alias{flat_class}
\alias{is_flexfile_flat}
\alias{is_quantityreport_flat}
\title{*_flat class utilities}
\usage{
is_flexfile_flat(x)

is_quantityreport_flat(x)
}
\arguments{
\item{x}{An object to check for type '*_flat'.}
}
\description{
Functions to help with with the 'flexfile_flat' and 'quantityreport_flat' class type.

\code{is_flexfile_flat()} checks if object is of type 'flexfile_flat'

\code{is_quantityreport_flat()} checks if object is of type 'quantityreport_flat'
}
