% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data.flexfile}
\alias{flatten_data.flexfile}
\title{Create a cost and hour dataframe from a FlexFile}
\usage{
\method{flatten_data}{flexfile}(x, .allocate = TRUE, ...)
}
\arguments{
\item{x}{A list of one or more collections of data frames to be flattened.}

\item{.allocate}{Logical whether or not to apply the allocations before flattening. In almost all
cases this should be left as \code{TRUE}.}

\item{...}{Arguments passed on to methods.}
}
\description{
\code{flatten_data.flexfile()} creates a single flat file for working with FlexFile
data. Input should be a list of one or more FlexFiles imported through the \code{read_flexfile} function.\cr
\cr
All fields retain their original names from the data model unless they are added in the flattening process. These
new fields are in 'snake_case' rather than ProperCase.
}
\examples{
\dontrun{
# Flatten one FlexFile
file <- system.file("extdata", "Sample_FlexFile_A.zip", package = "flexample")

flat_flex_file <- read_flexfile(file) \%>\%
  flatten_data()

# Flatten multiple FlexFiles
library(dplyr)

files <- system.file("extdata", package = "flexample")

flat_flexfiles <- read_folder(files, read_flexfile) \%>\%
  flatten_data()
  bind_rows(.id = "doc_id")
}
}
