% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data.list}
\alias{flatten_data.list}
\title{Flatten list of data}
\usage{
\method{flatten_data}{list}(x, ...)
}
\arguments{
\item{x}{A list of one or more collections of data frames to be flattened.}

\item{...}{Arguments passed on to methods.}
}
\description{
\code{flatten_data.list()} will check if each item in the list has the same class. If
so, iterate over each item and let its method dispatch.
}
