% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data.quantityreport}
\alias{flatten_data.quantityreport}
\title{Create a cost and hour dataframe from a Quantity Data Report}
\usage{
\method{flatten_data}{quantityreport}(x, ...)
}
\arguments{
\item{x}{A list of one or more collections of data frames to be flattened.}

\item{...}{Arguments passed on to methods.}
}
\description{
\code{flatten_data.quantityreport()} creates a single flat file for working with
the Quantity Data Report. Input should be a list of one or more Quantity Data Reports
imported through the \code{read_flexfile} function.
}
\examples{
\dontrun{
# Flatten one Quantity Report
file <- system.file("extdata", "Sample_Quantity_A.zip", package = "flexample")

flat_flex_file <- read_flexfile(file) \%>\%
  flatten_data()
}
}
