% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data}
\alias{flatten_data}
\alias{flatten_data.list}
\alias{flatten_data.flexfile}
\alias{flatten_data.quantityreport}
\title{Create a flat file from multiple data frames}
\usage{
flatten_data(x, ...)

\method{flatten_data}{list}(x, ...)

\method{flatten_data}{flexfile}(x, .allocate = TRUE, ...)

\method{flatten_data}{quantityreport}(x, ...)
}
\arguments{
\item{x}{A list of one or more collections of data frames to be flattened.}

\item{...}{Arguments passed on to methods.}

\item{.allocate}{Logical whether or not to apply the allocations before flattening. In almost all
cases this should be left as \code{TRUE}.}
}
\value{
A data frame with the flat file representation of the input data.
}
\description{
A generic function that is used to create a single flat file from a list of data
frames. This list is usually created by reading in a data format
with data spanning multiple tables.

\code{flatten_data.list()} will check if each item in the list has the same class. If
so, iterate over each item and let its method dispatch.

\code{flatten_data.flexfile()} creates a single flat file for working with FlexFile
data. Input should be a list of one or more FlexFiles imported through the \code{read_flexfile} function.

\code{flatten_data.quantityreport()} creates a single flat file for working with
the Quantity Data Report. Input should be a list of one or more Quantity Data Reports
imported through the \code{read_flexfile} function.
}
\examples{
\dontrun{
# Flatten one FlexFile
file <- system.file("extdata", "Sample_FlexFile_A.zip", package = "flexample")

flat_flex_file <- read_flexfile(file) \%>\%
  flatten_data()

# Flatten multiple FlexFiles
library(dplyr)

files <- system.file("extdata", package = "flexample")

flat_flexfiles <- read_folder(files, read_flexfile) \%>\%
  flatten_data()
  bind_rows(.id = "doc_id")
}
\dontrun{
# Flatten one Quantity Report
file <- system.file("extdata", "Sample_Quantity_A.zip", package = "flexample")

flat_flex_file <- read_flexfile(file) \%>\%
  flatten_data()
}
}
