% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_ff}
\alias{flatten_ff}
\title{Flatten a FlexFile report}
\usage{
flatten_ff(
  file,
  .show_check = FALSE,
  .coerce_spec = TRUE,
  .warn_utf8_bom = TRUE
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{flatten_ff()} was removed in favor of \code{\link{flatten_data}()}. See
\code{vignette("importing-flexfile")} for the preferred workflow.
}
\keyword{internal}
