% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexfile_enum.R
\docType{data}
\name{flexfile_enum}
\alias{flexfile_enum}
\title{FlexFile enumeration tables}
\format{
An object of class \code{list} of length 9.
}
\usage{
flexfile_enum
}
\description{
A list containing nine data frames: \code{phaseormilestoneenum}, \code{contracttypeenum},
\code{appropriationtypeenum}, \code{reportcycleenum}, \code{nonrecurringorrecurringenum},
\code{standardcategoryenum}, \code{detailedstandardcategoryenum}, \code{allocationmethodtypeenum}, and
\code{costhourtagenum}.
}
\keyword{datasets}
