% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexfile_enum_spec.R
\docType{data}
\name{flexfile_enum_spec}
\alias{flexfile_enum_spec}
\title{File specification for the FlexFile}
\format{
An object of class \code{list} of length 2.
}
\usage{
flexfile_enum_spec
}
\description{
A list containing two data frames: \code{fields} and \code{tables}.
}
\section{tables}{

\code{tables} is a data frame with 9 rows and 5 variables:
\describe{
\item{table}{The table name in the specification}
\item{entity}{The entity name in the specification}
\item{snake_table}{A snake_case version of table}
\item{type}{Either submission or enum}
\item{is_scalar}{Boolean whether the table is a list of scalars or a data table}
}
}

\section{fields}{

\code{fields} is a data frame with 18 rows and 9 variables:
\describe{
\item{table}{The table name in the specification}
\item{field}{The field name in the specification}
\item{snake_name}{A snake_case version of field}
\item{safe_name}{A database-safe version of field}
\item{type}{SQL data type for the field}
\item{pk}{Boolean if the field is a primary key}
\item{fk_table}{Foreign key constraint table, if applicable}
\item{fk_field}{Foreign key constraint field, if applicable}
\item{optional}{Boolean if the field is optional or not}
}
}

\seealso{
Other Data Specs: 
\code{\link{flexfile_spec}},
\code{\link{quantity_spec}}
}
\concept{Data Specs}
\keyword{datasets}
