% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flexfile_key_columns}
\alias{flexfile_key_columns}
\title{Key columns for the flexfile}
\usage{
flexfile_key_columns()
}
\description{
\code{flexfile_key_columns()} returns a vector of column names that uniquely define
a FlexFile and Quantity Data Report.
}
