% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantityreport-class.R
\name{is_quantityreport_list}
\alias{is_quantityreport_list}
\title{Check is object is a list of quantity reports}
\usage{
is_quantityreport_list(x)
}
\description{
Check is object is a list of quantity reports
}
