% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintrepair-class.R
\name{maintrepair_class}
\alias{maintrepair_class}
\alias{is_maintrepair}
\alias{is_maintrepair_list}
\title{maintrepair class utilities}
\usage{
is_maintrepair(x)

is_maintrepair_list(x)
}
\arguments{
\item{x}{An object to test or coerce to type 'maintrepair'.}
}
\description{
Functions to help with with the 'maintrepair' class type.

\code{is_maintrepair()} checks if object is of type 'maintrepair'

\code{is_maintrepair_list()} check if the object is a list where all members are of class 'maintrepair'.
}
