% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantityreport-class.R
\name{quantityreport_class}
\alias{quantityreport_class}
\alias{is_quantityreport}
\alias{as_quantityreport}
\alias{is_quantityreport_list}
\title{quantityreport class utilities}
\usage{
is_quantityreport(x)

as_quantityreport(
  x,
  names_case = c("snake_case", "data_model"),
  .drop_optional = TRUE,
  .show_check = TRUE
)

is_quantityreport_list(x)
}
\arguments{
\item{x}{An object to test or coerce to type 'quantityreport'.}

\item{names_case}{Case of the object being passed in.}

\item{.drop_optional}{Logical whether to drop optional columns or not. Not recommended for use because
it can result in downstream data inconsistencies.}

\item{.show_check}{Logical whether or not to show results from the check against
the file specification.}
}
\description{
Functions to help with with the 'quantityreport' class type.

\code{is_quantityreport()} checks if object is of type 'quantityreport'.

\code{as_quantityreport()} creates object of type 'quantityreport' and checks it against
the file specification.

\code{is_quantityreport_list()} check if the object is a list where all members are of class 'quantityreport'.
}
