% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_template.R
\name{read_excel_template}
\alias{read_excel_template}
\title{Read Maintenance & Repair Parts report}
\usage{
read_excel_template(
  file,
  table_spec,
  file_type,
  constructor,
  .show_check = FALSE,
  .coerce_spec = TRUE
)
}
\arguments{
\item{file}{Path to a Maintenance & Repair Part Report or Technical Data Report Excel file.}

\item{table_spec}{List containing data specifications for both tables, and fields within the tables}

\item{file_type}{Type of excel template used (e.g. "M&R Report", "TDR Report"). Used to inform check_spec function.}

\item{constructor}{Constructor to create object of the class pertaining to the file type (e.g. new_maintrepair, new_techdatareport)}

\item{.show_check}{Logical whether to print information about the file check to the console or not.}

\item{.coerce_spec}{Logical whether to coerce all column data types to those from the data models.
If \code{FALSE}, the types will be as detected upon read by the Excel reader.}
}
\value{
A list of tibbles for the \code{file}. Result will be of class \code{maintrepair} or \code{techdatareport}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{read_excel_template()} returns a list of tibbles from an Excel template. Templates currently supported include
the Maintenance and Repair Parts Report (MNR), and the Technical Data Report (TDR). Each tibble corresponds to one Excel sheet.\cr
\cr
This function is currently experimental because it reads from the Excel template, which
may not be stable.
}
\seealso{
\link{maintrepair_class}

\link{techdatareport_class}
}
