% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readflexfile.R
\docType{package}
\name{readflexfile-package}
\alias{readflexfile}
\alias{readflexfile-package}
\title{readflexfile: Read FlexFile data}
\description{
To learn more about readflexfile, start with the vignettes:
\code{browseVignettes(package = "readflexfile")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Technomics/readflexfile}
}

}
\author{
\strong{Maintainer}: Adam H. James \email{ajames@technomics.net}

Authors:
\itemize{
  \item Ben Berkman \email{bberkman@technomics.net}
  \item Justin Cooper \email{jcooper@technomics.net}
}

Other contributors:
\itemize{
  \item Technomics, Inc [copyright holder]
}

}
\keyword{internal}
