% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_flexfile.R
\name{write_flexfile}
\alias{write_flexfile}
\title{Read FlexFile or Quantity report}
\usage{
write_flexfile(x, file)
}
\arguments{
\item{x}{An object of class \link{flexfile_class} or \link{quantityreport_class}.}

\item{file}{Path to a FlexFile or Quantity Report .zip archive. For the 3-Part Template,
a character vector of the three Excel files to read.}
}
\description{
\code{write_flexfile()} writes a list of tibbles into a zip folder of JSON files.
Each tibble corresponds to its respective JSON table. This function can write both a FlexFile
and a Quantity report.
}
