% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_ff.R
\name{flatten_lists}
\alias{flatten_lists}
\alias{flatten_ff}
\alias{flatten_qdr}
\title{Flatten FlexFile and Quantity Data Report}
\usage{
flatten_ff(flexfile)

flatten_qdr(quantity_data)
}
\arguments{
\item{flexfile}{A flexfile created from the \code{\link{read_ff}} function}

\item{quantity_data}{A quantity data report created from the \code{read_ff} function}
}
\value{
A data frame with the flattened flexfile or quantity data.
}
\description{
Create separate data frames for the FlexFile and the Quantity Data Report.
These functions merge together many tables into a single one more useful for analysis.
This generally involves joining together numerous ID fields. Actuals and forecasts are also
stacked into a single table.

\code{flatten_ff()} creates a master dataframe for working with FlexFile data.

\code{flatten_qdr()} creates a master dataframe for working with FlexFile quantity data.
}
\examples{
# read a sample flex file
file <- system.file("extdata", "Sample File_FF.zip", package = "csdrtools")

flat_flex_file <- read_ff(file) \%>\%
  costmisc::add_id_col() \%>\%
  flatten_ff()

# read a sample quantity report
file <- system.file("extdata", "Sample File_Q.zip", package = "csdrtools")

flat_flex_file <- read_ff(file) \%>\%
  costmisc::add_id_col() \%>\%
  flatten_qdr()

}
