% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexfile-class.R
\name{flexfile_class}
\alias{flexfile_class}
\alias{is_flexfile}
\alias{as_flexfile}
\alias{is_flexfile_list}
\title{flexfile class utilities}
\usage{
is_flexfile(x)

as_flexfile(x, allocated = FALSE, rolledup = FALSE, .show_check = TRUE)

is_flexfile_list(x)
}
\arguments{
\item{x}{An object to test or coerce to type 'flexfile'.}

\item{.show_check}{Logical whether or not to show results from the check against
the file specification}
}
\description{
Functions to help with with the 'flexfile' class type.

\code{is_flexfile()} checks if object is of type 'flexfile'

\code{as_flexfile()} creates object of type 'flexfile' and checks it against
the file specification.

\code{is_flexfile_list()} check if the object is a list where all members are of class 'flexfile'.
}
