% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexfile_spec.R
\docType{data}
\name{flexfile_spec}
\alias{flexfile_spec}
\title{File specification for the FlexFile}
\format{
An object of class \code{list} of length 2.
}
\usage{
flexfile_spec
}
\description{
A list containing two data frames: \code{fields} and \code{tables}.
}
\section{tables}{

\code{tables} is a data frame with 29 rows and 4 variables:
\describe{
\item{table}{The table name in the specification}
\item{entity}{The entity name in the specification}
\item{snake_table}{A snake_case version of table}
\item{type}{Either submission or enumeration}
}
}

\section{fields}{

\code{fields} is a data frame with 171 rows and 9 variables:
\describe{
\item{table}{The table name in the specification}
\item{field}{The field name in the specification}
\item{snake_name}{A snake_case version of field}
\item{safe_name}{A database-safe version of field}
\item{type}{SQL data type for the field}
\item{pk}{Boolean if the field is a primary key}
\item{fk_table}{Foreign key constraint table, if applicable}
\item{fk_field}{Foreign key constraint field, if applicable}
\item{optional}{Boolean if the field is optional or not}
}
}

\keyword{datasets}
