% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexfile_family.R
\name{identical_report_metadata}
\alias{identical_report_metadata}
\title{Check if Reports have Identical Metadata}
\usage{
identical_report_metadata(..., report_list = NULL)
}
\arguments{
\item{...}{Two or more reports from the FlexFile Family to check.}

\item{report_list}{Alternatively a list of reports to check. If specified, then
\code{...} will be ignored.}
}
\value{
Logical whether all of the reports are from the FlexFile family and with identical
Metadata tables.
}
\description{
\code{identical_report_metadata()} checks a collection reports to see if they have identical metadata.
This check will exclude the "PointOfContact" fields.
}
