% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ff.R
\name{read_ff}
\alias{read_ff}
\title{Read one FlexFile into a list of tibbles}
\usage{
read_ff(file, .clean_names = TRUE)
}
\arguments{
\item{file}{Path to a FlexFile .zip archive.}

\item{.clean_names}{Logical whether to clean the names using \code{\link[janitor]{clean_names}()}
or not.}
}
\value{
A list of tibbles for the \code{file}.
}
\description{
\code{read_ff()} returns a list of tibbles from a zip folder submission of the FlexFiles.
Each tibble corresponds to its respective JSON table.
}
\examples{
file <- system.file("extdata", "Sample File_FF.zip", package = "readflexfile")

flexfile <- read_ff(file)

}
