% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flexfile.R
\name{read_ff}
\alias{read_ff}
\title{Read FlexFile or Quantity report}
\usage{
read_ff(file, .show_check = FALSE, .coerce_spec = TRUE, .warn_utf8_bom = TRUE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{read_ff()} was renamed to \code{\link{read_flexfile}()} to be less ambiguous.
}
\keyword{internal}
