% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flexfile.R
\name{read_flexfile}
\alias{read_flexfile}
\title{Read FlexFile or Quantity report}
\usage{
read_flexfile(
  file,
  .show_check = FALSE,
  .coerce_spec = TRUE,
  .warn_utf8_bom = TRUE
)
}
\arguments{
\item{file}{Path to a FlexFile or Quantity Report .zip archive.}

\item{.show_check}{Logical whether to print information about the file check to the console or not.}

\item{.coerce_spec}{Logical whether to coerce all column data types to those from the data models.
If \code{FALSE}, the types will be as detected upon read by the JSON parser.}

\item{.warn_utf8_bom}{Logical whether to report UTF8 byte-order-mark errors.}
}
\value{
A list of tibbles for the \code{file}. Result will be either of class \code{flexfile} or
of class \code{quantityreport}.
}
\description{
\code{read_flexfile()} returns a list of tibbles from a zip folder submission of the FlexFiles.
Each tibble corresponds to its respective JSON table. This function can read both a FlexFile
and a Quantity report.
}
\details{
Can be used with \code{\link[costmisc:read_folder]{read_folder}} as in example.
}
\examples{
\dontrun{
# Read in one FlexFile
file <- system.file("extdata", "Sample_FlexFile_A.zip", package = "flexample")

flexfile <- read_flexfile(file)

# Read in multiple FlexFiles by using read_folder
folder <- system.file("extdata", package = "flexample")

flexfiles <- read_folder(folder, read_flexfile)
}
}
\seealso{
\link{flexfile_class}, \link{quantityreport_class}
}
