% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maintrepair.R
\name{read_maintrepair}
\alias{read_maintrepair}
\title{Read Maintenance & Repair Parts report}
\usage{
read_maintrepair(file, .show_check = FALSE, .coerce_spec = TRUE)
}
\arguments{
\item{file}{Path to a Maintenance & Repair Part report Excel file.}

\item{.show_check}{Logical whether to print information about the file check to the console or not.}

\item{.coerce_spec}{Logical whether to coerce all column data types to those from the data models.
If \code{FALSE}, the types will be as detected upon read by the Excel reader.}
}
\value{
A list of tibbles for the \code{file}. Result will be either of class \code{maintrepair}.
}
\description{
\code{read_maintrepair()} returns a list of tibbles from an Excel submission of the Maintenance
and Repair Parts report. Each tibble corresponds to its Excel sheet.
}
\details{
Can be used with \code{\link[costmisc:read_folder]{read_folder}} as in example.
}
\seealso{
\link{maintrepair_class}
}
