% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_mapping.R
\docType{data}
\name{sfc_mapping}
\alias{sfc_mapping}
\title{Standard functional category mapping table}
\format{
\code{sfc_mapping} is a data frame with 26 rows and 4 variables:
\describe{
\item{standard_category_id}{The standard category identifier}
\item{standard_category}{The standard category}
\item{detailed_standard_category_id}{The detailed category identifier}
\item{ direct_or_overhead}{Either "direct" or "overhead"}
}
}
\usage{
sfc_mapping
}
\description{
A data frame containing the mapping between standard and detailed categories.
}
\keyword{datasets}
