% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ff.R
\name{stack_ff}
\alias{stack_ff}
\title{Stack list of FlexFiles or Quantity Reports into one list}
\usage{
stack_ff(data)
}
\arguments{
\item{data}{A list of FlexFile or Quantity Report tibbles converted from JSON format.}
}
\value{
A list of stacked tibbles.
}
\description{
\code{stack_ff()} reads in a list of lists of FlexFile or Quantity Report tibbles and returns a
single list of stacked tibbles. The \code{data} list should not mix report type (i.e.,
it should contain all FlexFiles or all Quantity Reports).\cr
\cr
This is a thin wrapper around \code{\link[costmisc]{unnest_df}()}.
}
\examples{
\dontrun{
files <- system.file("extdata", package = "flexample")

flexfiles <- read_folder(files, read_ff) \%>\%
  listindex_to_col() \%>\%
  stack_ff()
}
}
