% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ff.R
\name{stack_ff}
\alias{stack_ff}
\title{Stack list of multiple FlexFile submissions into one list of tibbles}
\usage{
stack_ff(.data)
}
\arguments{
\item{.data}{A list of FlexFile submissions' tibbles converted from JSON format.}
}
\value{
A list of stacked tibbles of multiple dataframes
}
\description{
\code{stack_ff()} reads in a list of lists of FlexFile submissions returns one list of stacked tibbles.
Thin wrapper around \code{\link{unnest_df}()}.
}
\examples{
\dontrun{
files <- system.file("extdata", package = "readflexfile")

flexfiles <- read_folder(files, read_ff) \%>\%
listindex_to_col() \%>\%
stack_ff()
}
}
